Extract classes into separate files in lib
