+++ spec/integration/validation_spec.rb
@@ -1,91 +1,3 @@
-require 'byebug'
-
-module Dry
-  module Validation
-    def self.Result(input)
-      case input
-      when Result then input
-      else Result.new(input)
-      end
-    end
-
-    class Result
-      attr_reader :value
-
-      def initialize(value)
-        @value = value
-      end
-
-      def success?
-        @value
-      end
-
-      def failure?
-        ! success?
-      end
-
-      def &(other)
-        if failure?
-          self
-        else
-          other
-        end
-      end
-    end
-
-    class Predicate
-      attr_reader :fn
-
-      class Composite < Predicate
-      end
-
-      def initialize(&block)
-        @fn = block
-      end
-
-      def call(*args)
-        Validation.Result(fn.call(*args))
-      end
-
-      def &(other)
-        Predicate::Composite.new { |*args| Validation.Result(fn.call(*args) && fn.call(*args)) }
-      end
-
-      def inversed
-        self.class.new { |*args| Validation.Result(!fn.(*args)) }
-      end
-
-      def curry(*args)
-        self.class.new(&fn.curry.(*args))
-      end
-    end
-
-    class Schema
-      extend Dry::Configurable
-
-      setting :predicates
-      setting :rules, Hash.new { |k, v| k[v] = [] }
-
-      attr_reader :rules
-
-      def self.attribute(name, predicate)
-        config.rules[name] = config.predicates[predicate].curry(name)
-      end
-
-      def initialize
-        @rules = self.class.config.rules
-      end
-
-      def call(input)
-        rules.each_with_object(Hash.new { |k,v| k[v] = [] }) do |(name, rule), errors|
-          result = rule.(input)
-          errors[name] << rule if result.failure?
-        end
-      end
-    end
-  end
-end
-
 RSpec.describe Dry::Validation do
   before do
     module Test
